YYC-108: split tui/mod.rs into focused submodules
tui/mod.rs grew to ~2700 lines covering input handling, rendering,
slash commands, pickers, diff scrubber, terminal init/restore, and
the full streaming-event pipeline. Carve out the concrete subsystems
into siblings so reviewers can reason about a single surface at a
time.

* tui/events.rs — RenderWake + render_wake_for_stream_batch +
  stream_event_forces_redraw + handle_stream_event. Plus the frame
  pacing tests.
* tui/commands.rs — SlashCommand struct, SLASH_COMMANDS catalog,
  filter_commands / current_palette / complete_slash, and the
  format_model_list / format_provider_list / build_provider_picker_entries
  helpers + their tests.
* tui/render_overlays.rs — every draw_* helper (draw_palette,
  draw_session_picker, draw_model_picker, draw_provider_picker,
  draw_diff_scrubber, draw_picker_border) plus the shared primitives
  (render_picker_column, build_picker_details, trim_to_width,
  picker_current_leaf).
* tui/picker_actions.rs — picker_move / picker_drill_or_commit /
  picker_commit_current / initial_path_for_active_model. State
  mutations only — no rendering.
* tui/init.rs — init_terminal / restore_terminal lifecycle helpers.

mod.rs keeps run_tui (the agent + event loop wiring), the
ResumeTarget public enum, KeyEv, short_id, submit_prompt, and
refresh_sessions. Public surface unchanged.

cargo test --all-features --lib passes (252 tests).
cargo clippy --all-features --all-targets -- -D warnings still clean.

Note on size: the issue's <1000 target on tui/mod.rs would require
splitting run_tui itself (currently ~1290 lines covering keybinds
dispatch, slash command handlers, picker overlays, pause routing,
and the redraw scheduler). That's a deeper refactor than the
structural split scoped here; mod.rs lands at 1412 lines. Tracking
the run_tui body split as a follow-up.
