+++ src/tui/mod.rs
@@ -1,15 +1,11 @@
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use anyhow::Result;
 use ratatui::{
-    Terminal,
     crossterm::event::{self, Event, KeyCode, KeyEventKind, KeyModifiers},
     layout::Rect,
     prelude::Position,
-    style::{Modifier, Style},
-    text::{Line, Span},
-    widgets::{Paragraph, Wrap},
 };
 use tokio::sync::{Mutex, mpsc};
 
@@ -22,48 +18,38 @@ use crate::provider::StreamEvent;
 
 pub mod chat_message;
 pub mod chat_render;
+mod commands;
+mod events;
+mod init;
 pub mod keybinds;
 pub mod markdown;
 pub mod model_picker;
 pub mod orchestration;
+mod picker_actions;
 pub mod picker_state;
+mod render_overlays;
 pub mod state;
 pub mod theme;
 pub mod views;
 pub mod widgets;
 
-use state::{AppState, ChatMessage, ChatRole, SessionStatus};
-use theme::{Theme, body};
+use commands::{
+    SLASH_COMMANDS, build_provider_picker_entries, complete_slash, current_palette,
+    filter_commands,
+};
+use events::{
+    RenderWake, handle_stream_event, render_wake_for_stream_batch, stream_event_forces_redraw,
+};
+use init::{init_terminal, restore_terminal};
+use picker_actions::{
+    initial_path_for_active_model, picker_commit_current, picker_drill_or_commit, picker_move,
+};
+use render_overlays::{
+    draw_diff_scrubber, draw_model_picker, draw_palette, draw_provider_picker, draw_session_picker,
+};
+use state::{AppState, ChatMessage, ChatRole};
+use theme::Theme;
 use views::{View, render_view};
-
-const STREAM_FRAME_BUDGET: Duration = Duration::from_millis(16);
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum RenderWake {
-    Now,
-    Wait(Duration),
-}
-
-fn render_wake_for_stream_batch(
-    last_draw: Instant,
-    now: Instant,
-    is_terminal_event: bool,
-) -> RenderWake {
-    if is_terminal_event {
-        return RenderWake::Now;
-    }
-
-    let elapsed = now.saturating_duration_since(last_draw);
-    if elapsed >= STREAM_FRAME_BUDGET {
-        RenderWake::Now
-    } else {
-        RenderWake::Wait(STREAM_FRAME_BUDGET - elapsed)
-    }
-}
-
-fn stream_event_forces_redraw(ev: &StreamEvent) -> bool {
-    matches!(ev, StreamEvent::Done(_) | StreamEvent::Error(_))
-}
 
 /// What session, if any, the TUI should load on startup.
 #[derive(Debug, Clone)]
@@ -85,225 +71,8 @@ enum KeyEv {
     Error(String),
 }
 
-#[derive(Debug, Clone)]
-struct SlashCommand {
-    name: &'static str,
-    description: &'static str,
-    /// True when the command can run mid-turn without corrupting agent state
-    /// (YYC-62). Pure UI ops are safe; anything that mutates conversation
-    /// history or reaches into the agent is not. Default false (conservative).
-    mid_turn_safe: bool,
-}
-
-const SLASH_COMMANDS: &[SlashCommand] = &[
-    SlashCommand {
-        name: "exit",
-        description: "Quit Vulcan",
-        // Always exits cleanly; no state to corrupt.
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "quit",
-        description: "Quit Vulcan",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "help",
-        description: "Show available commands",
-        // Pure UI: pushes a system message.
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "clear",
-        description: "Clear message history",
-        // Destructive: would nuke the in-flight User+Agent pair the agent
-        // loop is streaming into. Defer until idle.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "view",
-        description: "Cycle to next view (or 1-5)",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "reasoning",
-        description: "Toggle reasoning trace",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "search",
-        description: "Search past sessions: /search <query>",
-        // Holds agent.lock().await — would deadlock against the in-flight
-        // run_prompt_stream task. Defer until idle.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "model",
-        description: "List or switch models: /model [id]",
-        // Rebuilds the provider for future turns and may fetch the catalog.
-        // Defer until idle so the in-flight provider stream is untouched.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "provider",
-        description: "List or switch named providers: /provider [name|default]",
-        // Rebuilds the provider against a different profile; same idle
-        // requirement as /model.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "diff-style",
-        description: "Set diff render: /diff-style <unified|side-by-side|inline>",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "resume",
-        description: "Open session picker to switch to another session",
-        mid_turn_safe: false,
-    },
-];
-
-fn short_id(id: &str) -> String {
+pub(super) fn short_id(id: &str) -> String {
     id.chars().take(8).collect()
-}
-
-#[allow(dead_code)] // retained for tests and potential `--help`-style printers.
-fn format_model_list(active_model: &str, models: &[crate::provider::catalog::ModelInfo]) -> String {
-    let mut out = format!("Models from active provider ({} total):", models.len());
-    for model in models.iter().take(30) {
-        let marker = if model.id == active_model { "*" } else { " " };
-        let context = if model.context_length > 0 {
-            crate::tui::state::format_thousands(model.context_length as u32)
-        } else {
-            "unknown".into()
-        };
-        let mut flags = Vec::new();
-        if model.features.tools {
-            flags.push("tools");
-        }
-        if model.features.reasoning {
-            flags.push("reasoning");
-        }
-        if model.features.vision {
-            flags.push("vision");
-        }
-        if model.features.json_mode {
-            flags.push("json");
-        }
-        let flags = if flags.is_empty() {
-            String::new()
-        } else {
-            format!(" · {}", flags.join(","))
-        };
-        out.push_str(&format!("\n  {marker} {} · ctx {context}{flags}", model.id));
-    }
-    if models.len() > 30 {
-        out.push_str(&format!("\n  ... {} more", models.len() - 30));
-    }
-    out.push_str("\n\nUse /model <id> to switch.");
-    out
-}
-
-fn build_provider_picker_entries(config: &Config) -> Vec<crate::tui::state::ProviderPickerEntry> {
-    use crate::tui::state::ProviderPickerEntry;
-    let mut out = Vec::with_capacity(config.providers.len() + 1);
-    out.push(ProviderPickerEntry {
-        name: None,
-        model: config.provider.model.clone(),
-        base_url: config.provider.base_url.clone(),
-    });
-    let mut names: Vec<&String> = config.providers.keys().collect();
-    names.sort();
-    for name in names {
-        let cfg = &config.providers[name];
-        out.push(ProviderPickerEntry {
-            name: Some(name.clone()),
-            model: cfg.model.clone(),
-            base_url: cfg.base_url.clone(),
-        });
-    }
-    out
-}
-
-#[allow(dead_code)] // retained for tests and potential `--help`-style printers.
-fn format_provider_list(config: &Config, active: Option<&str>) -> String {
-    let mut out = String::from("Provider profiles:");
-    let default_marker = if active.is_none() { "*" } else { " " };
-    out.push_str(&format!(
-        "\n  {default_marker} default · {} · {}",
-        config.provider.base_url, config.provider.model,
-    ));
-
-    let mut names: Vec<&String> = config.providers.keys().collect();
-    names.sort();
-    for name in names {
-        let cfg = &config.providers[name];
-        let marker = if active == Some(name.as_str()) {
-            "*"
-        } else {
-            " "
-        };
-        out.push_str(&format!(
-            "\n  {marker} {name} · {} · {}",
-            cfg.base_url, cfg.model,
-        ));
-    }
-    if config.providers.is_empty() {
-        out.push_str("\n  (no named [providers.<name>] profiles configured)");
-    }
-    out.push_str("\n\nUse /provider <name> to switch, /provider default to revert.");
-    out
-}
-
-fn filter_commands(prefix: &str) -> Vec<&'static SlashCommand> {
-    if prefix.is_empty() {
-        return SLASH_COMMANDS.iter().collect();
-    }
-    let lower = prefix.to_lowercase();
-    SLASH_COMMANDS
-        .iter()
-        .filter(|c| c.name.starts_with(&lower))
-        .collect()
-}
-
-/// Same matching logic as the palette renderer in the main loop — exposed
-/// as a helper so the key handler can decide what to highlight or commit
-/// without duplicating prefix logic (YYC-70).
-fn current_palette(input: &str) -> Vec<&'static SlashCommand> {
-    if input == "/" {
-        SLASH_COMMANDS.iter().collect()
-    } else if input.starts_with('/') && input.len() > 1 {
-        filter_commands(&input[1..])
-    } else {
-        Vec::new()
-    }
-}
-
-fn complete_slash(prefix: &str) -> Option<String> {
-    let matches = filter_commands(prefix);
-    if matches.is_empty() {
-        return None;
-    }
-    if matches.len() == 1 {
-        return Some(matches[0].name.to_string());
-    }
-    let first = matches[0].name.as_bytes();
-    let mut common = first.len();
-    for m in &matches[1..] {
-        let bytes = m.name.as_bytes();
-        common = common.min(bytes.len());
-        for (i, &b) in first.iter().enumerate().take(common) {
-            if b != bytes[i] {
-                common = i;
-                break;
-            }
-        }
-    }
-    if common > prefix.len() {
-        Some(matches[0].name[..common].to_string())
-    } else {
-        None
-    }
 }
 
 /// Spawn a fresh agent turn for `msg`. Updates chat state (User + empty
@@ -350,110 +119,6 @@ async fn refresh_sessions(agent: &Arc<Mutex<Agent>>, app: &mut AppState) {
     app.hydrate_sessions(&summaries, &active_session_id);
 }
 
-async fn handle_stream_event(
-    app: &mut AppState,
-    agent: &Arc<Mutex<Agent>>,
-    stream_tx: &mpsc::UnboundedSender<StreamEvent>,
-    ev: StreamEvent,
-) {
-    match ev {
-        StreamEvent::Text(chunk) => {
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent) {
-                    // Append to both the segment timeline (the YYC-71
-                    // ordered renderer) and the legacy `content` field
-                    // (kept so other code that peeks at .content keeps working).
-                    last.append_text(&chunk);
-                    last.content.push_str(&chunk);
-                }
-        }
-        StreamEvent::Reasoning(chunk) => {
-            // Per-token reasoning trace from thinking-mode models. Push to
-            // the segment timeline so it interleaves with tool calls in
-            // render order; also append to the legacy `reasoning` field so
-            // latest_reasoning() etc. continue to work.
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent) {
-                    last.append_reasoning(&chunk);
-                    last.reasoning.push_str(&chunk);
-                }
-            app.note_reasoning();
-        }
-        StreamEvent::Done(resp) => {
-            app.thinking = false;
-            if let Some(usage) = resp.usage {
-                // YYC-60: track lifetime totals for cost (YYC-67) and the
-                // latest prompt size for the in-status capacity bar.
-                app.prompt_tokens_total = app
-                    .prompt_tokens_total
-                    .saturating_add(usage.prompt_tokens as u32);
-                app.completion_tokens_total = app
-                    .completion_tokens_total
-                    .saturating_add(usage.completion_tokens as u32);
-                app.prompt_tokens_last = usage.prompt_tokens as u32;
-            }
-            app.note_done();
-            refresh_sessions(agent, app).await;
-            // YYC-61: drain one queued prompt per turn end. Subsequent queued
-            // prompts ride the next Done event in the same way.
-            if let Some(next) = app.queue.pop_front() {
-                submit_prompt(app, agent, stream_tx, next);
-            }
-        }
-        StreamEvent::Error(e) => {
-            if let Some(last) = app.messages.last_mut()
-                && last.content.is_empty() {
-                    last.set_content(format!("⚠ Error: {e}"));
-                }
-            app.thinking = false;
-            // YYC-67: record provider-level error for telemetry.
-            app.provider_errors_total = app.provider_errors_total.saturating_add(1);
-            app.note_error(&e);
-        }
-        StreamEvent::ToolCallStart {
-            name, args_summary, ..
-        } => {
-            // YYC-71: push the tool-call segment into the timeline
-            // (interleaved with reasoning/text). YYC-74: carry the args
-            // summary so the card has structured context.
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent) {
-                    last.push_tool_start_with(name.clone(), args_summary);
-                }
-            app.note_tool_start(&name);
-        }
-        StreamEvent::ToolCallEnd {
-            name,
-            ok,
-            output_preview,
-            result_meta,
-            elided_lines,
-            elapsed_ms,
-            ..
-        } => {
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent) {
-                    // YYC-74: stamp preview + meta + timing onto the matching
-                    // segment for the card. YYC-78: stash elided count for the
-                    // collapse footer.
-                    last.finish_tool_with(
-                        &name,
-                        ok,
-                        output_preview,
-                        result_meta,
-                        elided_lines,
-                        Some(elapsed_ms),
-                    );
-                }
-            // YYC-67: tool call telemetry.
-            app.tool_calls_total = app.tool_calls_total.saturating_add(1);
-            if !ok {
-                app.tool_errors_total = app.tool_errors_total.saturating_add(1);
-            }
-            app.note_tool_end(&name, ok);
-        }
-    }
-}
 
 pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
     let mut terminal = init_terminal()?;
@@ -1745,941 +1410,3 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
     Ok(())
 }
 
-fn draw_palette(
-    f: &mut ratatui::Frame,
-    area: Rect,
-    cmds: &[&SlashCommand],
-    selected: usize,
-    theme: &Theme,
-) {
-    if area.height == 0 {
-        return;
-    }
-    // Title bar
-    let bar = Rect {
-        x: area.x,
-        y: area.y,
-        width: area.width,
-        height: 1,
-    };
-    let mut header_text = " ▓▓ COMMANDS".to_string();
-    if (header_text.chars().count() as u16) < bar.width {
-        header_text.push_str(&" ".repeat(bar.width as usize - header_text.chars().count()));
-    }
-    f.render_widget(
-        Paragraph::new(header_text).style(theme.accent.add_modifier(Modifier::BOLD)),
-        bar,
-    );
-    let inner = Rect {
-        x: area.x,
-        y: area.y + 1,
-        width: area.width,
-        height: area.height.saturating_sub(1),
-    };
-    let mut lines = Vec::new();
-    if cmds.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "  No matching commands",
-            theme.muted,
-        )));
-    } else {
-        let active = selected.min(cmds.len().saturating_sub(1));
-        for (i, cmd) in cmds.iter().enumerate() {
-            let is_active = i == active;
-            // YYC-70: highlight the active row by swapping fg/bg of accent
-            // (gives a visible selection bar regardless of active theme).
-            let (prefix, name_style, desc_style) = if is_active {
-                let active_style = theme.accent.add_modifier(Modifier::BOLD);
-                ("▸ ", active_style, active_style)
-            } else {
-                (
-                    "  ",
-                    theme.accent.add_modifier(Modifier::BOLD),
-                    theme.assistant,
-                )
-            };
-            lines.push(Line::from(vec![
-                Span::styled(format!("{prefix}/{:<12}", cmd.name), name_style),
-                Span::styled(cmd.description, desc_style),
-            ]));
-        }
-    }
-    f.render_widget(Paragraph::new(lines).style(body()), inner);
-}
-
-fn draw_session_picker(f: &mut ratatui::Frame, area: Rect, app: &AppState) {
-    let theme = &app.theme;
-    let width = area.width.min(56);
-    let height = (app.sessions.len() as u16 + 6).min(area.height.saturating_sub(2));
-    let x = area.x + (area.width.saturating_sub(width)) / 2;
-    let y = area.y + (area.height.saturating_sub(height)) / 2;
-
-    let box_area = Rect {
-        x,
-        y,
-        width,
-        height,
-    };
-    if box_area.height < 4 {
-        return;
-    }
-
-    // Title bar
-    let bar = Rect {
-        x: box_area.x,
-        y: box_area.y,
-        width: box_area.width,
-        height: 1,
-    };
-    let mut title = "  Resume a Session  ".to_string();
-    if (title.chars().count() as u16) < bar.width {
-        let pad = bar.width as usize - title.chars().count();
-        title = format!(
-            "{}{}{}",
-            " ".repeat(pad / 2),
-            title.trim(),
-            " ".repeat(pad - pad / 2)
-        );
-    }
-    f.render_widget(
-        Paragraph::new(title).style(theme.accent.add_modifier(Modifier::BOLD)),
-        bar,
-    );
-
-    // Session list
-    let list_area = Rect {
-        x: box_area.x,
-        y: box_area.y + 1,
-        width: box_area.width,
-        height: box_area.height.saturating_sub(2),
-    };
-    let mut lines: Vec<Line<'static>> = Vec::new();
-
-    if app.sessions.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "  No saved sessions found.",
-            theme.muted,
-        )));
-        lines.push(Line::from(Span::styled(
-            "  Start a conversation and sessions will appear here.",
-            theme.muted,
-        )));
-    } else {
-        let active = app
-            .session_picker_selection
-            .min(app.sessions.len().saturating_sub(1));
-        for (i, s) in app.sessions.iter().enumerate() {
-            let is_active = i == active;
-            let marker = if is_active { "▸ " } else { "  " };
-            let status_style = match s.status {
-                SessionStatus::Live => theme.success,
-                SessionStatus::Saved => theme.system,
-            };
-            let status_label = match s.status {
-                SessionStatus::Live => "LIVE",
-                SessionStatus::Saved => "saved",
-            };
-
-            let dt = chrono::DateTime::from_timestamp(s.last_active, 0)
-                .map(|d| {
-                    d.with_timezone(&chrono::Local)
-                        .format("%b %d %H:%M")
-                        .to_string()
-                })
-                .unwrap_or_default();
-
-            let name_style = Style::default()
-                .fg(theme.body_fg)
-                .add_modifier(if is_active {
-                    Modifier::BOLD
-                } else {
-                    Modifier::empty()
-                });
-
-            lines.push(Line::from(vec![
-                Span::styled(marker, name_style.add_modifier(Modifier::BOLD)),
-                Span::styled("█ ", status_style),
-                Span::styled(format!("{:<12}", short_id(&s.id)), name_style),
-                Span::styled(format!("{:>4}m", s.message_count), theme.muted),
-                Span::styled(format!("  {} ", dt), theme.muted),
-                Span::styled(status_label, status_style.add_modifier(Modifier::BOLD)),
-            ]));
-
-            if let Some(preview) = &s.preview {
-                lines.push(Line::from(vec![
-                    Span::raw("  "),
-                    Span::styled(
-                        format!("  {}", preview),
-                        theme.muted.add_modifier(Modifier::DIM),
-                    ),
-                ]));
-            }
-        }
-    }
-
-    // Footer hint
-    let hint = "  ↑↓ navigate · Enter select · Esc cancel  ";
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled(hint, theme.muted)));
-
-    f.render_widget(
-        Paragraph::new(lines).style(Style::default().bg(theme.body_bg)),
-        list_area,
-    );
-
-    // Draw a border around the whole thing
-    draw_picker_border(f, box_area, theme);
-}
-
-fn draw_model_picker(f: &mut ratatui::Frame, area: Rect, app: &AppState) {
-    let theme = &app.theme;
-    let width = area.width.min(72);
-    let rows = (app.model_picker_items.len() as u16).min(20);
-    let height = (rows + 5).min(area.height.saturating_sub(2));
-    let x = area.x + (area.width.saturating_sub(width)) / 2;
-    let y = area.y + (area.height.saturating_sub(height)) / 2;
-    let box_area = Rect {
-        x,
-        y,
-        width,
-        height,
-    };
-    if box_area.height < 4 {
-        return;
-    }
-
-    let bar = Rect {
-        x: box_area.x,
-        y: box_area.y,
-        width: box_area.width,
-        height: 1,
-    };
-    let mut title = "  Switch Model  ".to_string();
-    if (title.chars().count() as u16) < bar.width {
-        let pad = bar.width as usize - title.chars().count();
-        title = format!(
-            "{}{}{}",
-            " ".repeat(pad / 2),
-            title.trim(),
-            " ".repeat(pad - pad / 2)
-        );
-    }
-    f.render_widget(
-        Paragraph::new(title).style(theme.accent.add_modifier(Modifier::BOLD)),
-        bar,
-    );
-
-    let list_area = Rect {
-        x: box_area.x,
-        y: box_area.y + 1,
-        width: box_area.width,
-        height: box_area.height.saturating_sub(2),
-    };
-
-    if app.model_picker_items.is_empty() {
-        f.render_widget(
-            Paragraph::new(Line::from(Span::styled(
-                "  No models in provider catalog.",
-                theme.muted,
-            ))),
-            list_area,
-        );
-        draw_picker_border(f, box_area, theme);
-        return;
-    }
-
-    // Mini-files-style miller columns. The number of visible columns is
-    // capped to whatever the box width can fit (each column is roughly
-    // 20–28 chars). Always reserve the rightmost column for details.
-    let drilled_depth = app.model_picker_path.len();
-    let max_tree_depth = app.model_picker_tree.max_depth();
-    let total_columns = max_tree_depth.max(1) + 1; // +1 for details
-
-    let col_width = (list_area.width / total_columns as u16).max(16);
-    let cols: Vec<Rect> = (0..total_columns)
-        .map(|i| Rect {
-            x: list_area.x + i as u16 * col_width,
-            y: list_area.y,
-            width: col_width,
-            height: list_area.height,
-        })
-        .collect();
-
-    // Render each tree column.
-    for (col_idx, col_rect) in cols.iter().enumerate().take(max_tree_depth) {
-        let path_prefix: Vec<usize> = app
-            .model_picker_path
-            .iter()
-            .copied()
-            .take(col_idx)
-            .collect();
-        let nodes = app.model_picker_tree.column_at(col_idx, &path_prefix);
-        let selection = app
-            .model_picker_path
-            .get(col_idx)
-            .copied()
-            .unwrap_or(0)
-            .min(nodes.len().saturating_sub(1));
-        let is_focused = col_idx == app.model_picker_focus;
-        render_picker_column(f, *col_rect, nodes, selection, is_focused, theme);
-    }
-
-    // Details panel at the rightmost column.
-    let details_col = cols.last().copied().unwrap_or(list_area);
-    let detail_lines = build_picker_details(app);
-    f.render_widget(
-        Paragraph::new(detail_lines).wrap(Wrap { trim: false }),
-        details_col,
-    );
-
-    let hint = "  hjkl move · Enter select · Esc cancel  (drilled: column ";
-    let footer_line = format!(
-        "{hint}{}/{})",
-        app.model_picker_focus + 1,
-        max_tree_depth.max(1)
-    );
-    let footer_rect = Rect {
-        x: list_area.x,
-        y: list_area.y + list_area.height.saturating_sub(1),
-        width: list_area.width,
-        height: 1,
-    };
-    f.render_widget(
-        Paragraph::new(Line::from(Span::styled(footer_line, theme.muted))),
-        footer_rect,
-    );
-    let _ = drilled_depth; // reserved for future filter UI
-
-    draw_picker_border(f, box_area, theme);
-}
-
-fn render_picker_column(
-    f: &mut ratatui::Frame,
-    area: Rect,
-    nodes: &[crate::tui::model_picker::TreeNode],
-    selection: usize,
-    is_focused: bool,
-    theme: &Theme,
-) {
-    let mut lines: Vec<Line<'static>> = Vec::new();
-    if nodes.is_empty() {
-        lines.push(Line::from(Span::styled("  ·", theme.muted)));
-    } else {
-        let visible = area.height.saturating_sub(2) as usize;
-        let start = selection.saturating_sub(visible.saturating_sub(1) / 2);
-        let end = (start + visible).min(nodes.len());
-        for (i, node) in nodes.iter().enumerate().take(end).skip(start) {
-            let is_active = i == selection;
-            let marker = if is_active && is_focused {
-                "▸ "
-            } else if is_active {
-                "│ "
-            } else {
-                "  "
-            };
-            let mut style = Style::default();
-            if is_active {
-                style = style.add_modifier(Modifier::BOLD);
-                if is_focused {
-                    style = if let Some(fg) = theme.accent.fg {
-                        style.fg(fg)
-                    } else {
-                        style.add_modifier(Modifier::REVERSED)
-                    };
-                }
-            }
-            let suffix = if node.children.is_empty() && node.model_index.is_some() {
-                ""
-            } else {
-                "›"
-            };
-            let label = trim_to_width(&node.label, area.width.saturating_sub(4) as usize);
-            lines.push(Line::from(vec![
-                Span::styled(marker, style),
-                Span::styled(label, style),
-                Span::styled(format!(" {suffix}"), theme.muted),
-            ]));
-        }
-        if start > 0 || end < nodes.len() {
-            lines.push(Line::from(Span::styled(
-                format!("  …{}/{}", end, nodes.len()),
-                theme.muted.add_modifier(Modifier::DIM),
-            )));
-        }
-    }
-    f.render_widget(Paragraph::new(lines), area);
-}
-
-fn build_picker_details(app: &AppState) -> Vec<Line<'static>> {
-    let theme = &app.theme;
-    let mut lines = Vec::new();
-    let leaf_idx = picker_current_leaf(&app.model_picker_tree, &app.model_picker_path);
-    let Some(idx) = leaf_idx else {
-        lines.push(Line::from(Span::styled(
-            "  drill in (l/→) for details",
-            theme.muted,
-        )));
-        return lines;
-    };
-    let Some(model) = app.model_picker_items.get(idx) else {
-        return lines;
-    };
-    lines.push(Line::from(Span::styled(
-        format!(" {}", model.id),
-        Style::default().add_modifier(Modifier::BOLD),
-    )));
-    lines.push(Line::from(""));
-    let ctx = if model.context_length > 0 {
-        crate::tui::state::format_thousands(model.context_length as u32)
-    } else {
-        "?".into()
-    };
-    lines.push(Line::from(Span::styled(
-        format!(" context  : {ctx}"),
-        theme.muted,
-    )));
-    let mut flags = Vec::new();
-    if model.features.tools {
-        flags.push("tools");
-    }
-    if model.features.reasoning {
-        flags.push("reasoning");
-    }
-    if model.features.vision {
-        flags.push("vision");
-    }
-    if model.features.json_mode {
-        flags.push("json");
-    }
-    let flag_str = if flags.is_empty() {
-        "(none reported)".to_string()
-    } else {
-        flags.join(", ")
-    };
-    lines.push(Line::from(Span::styled(
-        format!(" features : {flag_str}"),
-        theme.muted,
-    )));
-    if let Some(p) = &model.pricing {
-        lines.push(Line::from(Span::styled(
-            format!(
-                " pricing  : ${:.4}/1k in · ${:.4}/1k out",
-                p.input_per_token * 1000.0,
-                p.output_per_token * 1000.0,
-            ),
-            theme.muted,
-        )));
-    }
-    if let Some(top) = &model.top_provider {
-        lines.push(Line::from(Span::styled(
-            format!(" upstream : {top}"),
-            theme.muted,
-        )));
-    }
-    lines
-}
-
-fn trim_to_width(s: &str, width: usize) -> String {
-    if width == 0 {
-        return String::new();
-    }
-    let chars: Vec<char> = s.chars().collect();
-    if chars.len() <= width {
-        return s.to_string();
-    }
-    let head: String = chars.iter().take(width.saturating_sub(1)).collect();
-    format!("{head}…")
-}
-
-fn picker_move(app: &mut AppState, delta: i32) {
-    let depth = app.model_picker_focus;
-    let path_prefix: Vec<usize> = app.model_picker_path.iter().copied().take(depth).collect();
-    let len = app.model_picker_tree.column_at(depth, &path_prefix).len();
-    if len == 0 {
-        return;
-    }
-    while app.model_picker_path.len() <= depth {
-        app.model_picker_path.push(0);
-    }
-    let cur = app.model_picker_path[depth] as i32 + delta;
-    let max = (len - 1) as i32;
-    app.model_picker_path[depth] = cur.clamp(0, max) as usize;
-    // Reset deeper selections — the active branch changed.
-    app.model_picker_path.truncate(depth + 1);
-}
-
-fn picker_drill_or_commit(app: &mut AppState) -> Option<String> {
-    let depth = app.model_picker_focus;
-    let path_prefix: Vec<usize> = app.model_picker_path.iter().copied().take(depth).collect();
-    let nodes = app
-        .model_picker_tree
-        .column_at(depth, &path_prefix)
-        .to_vec();
-    let sel = app.model_picker_path.get(depth).copied().unwrap_or(0);
-    let node = nodes.get(sel)?;
-    if node.children.is_empty() {
-        // Leaf — commit.
-        return node
-            .model_index
-            .and_then(|i| app.model_picker_items.get(i))
-            .map(|m| m.id.clone());
-    }
-    // Drill: focus next column, default selection 0.
-    while app.model_picker_path.len() <= depth + 1 {
-        app.model_picker_path.push(0);
-    }
-    app.model_picker_path[depth + 1] = 0;
-    app.model_picker_focus = depth + 1;
-    None
-}
-
-fn picker_commit_current(app: &AppState) -> Option<String> {
-    picker_current_leaf(&app.model_picker_tree, &app.model_picker_path)
-        .and_then(|i| app.model_picker_items.get(i))
-        .map(|m| m.id.clone())
-}
-
-fn picker_current_leaf(
-    tree: &crate::tui::model_picker::ModelTree,
-    path: &[usize],
-) -> Option<usize> {
-    let mut current: &[crate::tui::model_picker::TreeNode] = &tree.labs;
-    let mut leaf: Option<usize> = None;
-    for &idx in path {
-        let node = current.get(idx)?;
-        if node.children.is_empty() {
-            return node.model_index;
-        }
-        leaf = node.model_index;
-        current = &node.children;
-    }
-    // Path didn't reach a leaf — return last seen leaf marker (None for
-    // internal-only nodes).
-    leaf
-}
-
-fn initial_path_for_active_model(
-    tree: &crate::tui::model_picker::ModelTree,
-    active_id: &str,
-    items: &[crate::provider::catalog::ModelInfo],
-) -> Vec<usize> {
-    let target = items.iter().position(|m| m.id == active_id);
-    fn find_path(
-        nodes: &[crate::tui::model_picker::TreeNode],
-        target: Option<usize>,
-        path: &mut Vec<usize>,
-    ) -> bool {
-        for (i, node) in nodes.iter().enumerate() {
-            path.push(i);
-            if node.model_index.is_some() && node.model_index == target {
-                return true;
-            }
-            if find_path(&node.children, target, path) {
-                return true;
-            }
-            path.pop();
-        }
-        false
-    }
-    let mut path = Vec::new();
-    if !find_path(&tree.labs, target, &mut path) {
-        // No exact match — start from column 0 with no drilled selection.
-        path.clear();
-        path.push(0);
-    }
-    path
-}
-
-fn draw_provider_picker(f: &mut ratatui::Frame, area: Rect, app: &AppState) {
-    let theme = &app.theme;
-    let width = area.width.min(72);
-    let rows = (app.provider_picker_items.len() as u16).min(12);
-    let height = (rows + 5).min(area.height.saturating_sub(2));
-    let x = area.x + (area.width.saturating_sub(width)) / 2;
-    let y = area.y + (area.height.saturating_sub(height)) / 2;
-    let box_area = Rect {
-        x,
-        y,
-        width,
-        height,
-    };
-    if box_area.height < 4 {
-        return;
-    }
-
-    let bar = Rect {
-        x: box_area.x,
-        y: box_area.y,
-        width: box_area.width,
-        height: 1,
-    };
-    let mut title = "  Switch Provider  ".to_string();
-    if (title.chars().count() as u16) < bar.width {
-        let pad = bar.width as usize - title.chars().count();
-        title = format!(
-            "{}{}{}",
-            " ".repeat(pad / 2),
-            title.trim(),
-            " ".repeat(pad - pad / 2)
-        );
-    }
-    f.render_widget(
-        Paragraph::new(title).style(theme.accent.add_modifier(Modifier::BOLD)),
-        bar,
-    );
-
-    let list_area = Rect {
-        x: box_area.x,
-        y: box_area.y + 1,
-        width: box_area.width,
-        height: box_area.height.saturating_sub(2),
-    };
-
-    let mut lines: Vec<Line<'static>> = Vec::new();
-    if app.provider_picker_items.is_empty() {
-        lines.push(Line::from(Span::styled("  (no providers)", theme.muted)));
-    } else {
-        let active = app
-            .provider_picker_selection
-            .min(app.provider_picker_items.len().saturating_sub(1));
-        for (i, e) in app.provider_picker_items.iter().enumerate() {
-            let is_active = i == active;
-            let marker = if is_active { "▸ " } else { "  " };
-            let label = e.name.clone().unwrap_or_else(|| "default".into());
-            let row_style = Style::default()
-                .fg(theme.body_fg)
-                .add_modifier(if is_active {
-                    Modifier::BOLD
-                } else {
-                    Modifier::empty()
-                });
-            lines.push(Line::from(vec![
-                Span::styled(marker, row_style.add_modifier(Modifier::BOLD)),
-                Span::styled(format!("{label:<12}"), row_style),
-                Span::styled(format!(" {}", e.model), theme.muted),
-                Span::styled(format!("  ({})", e.base_url), theme.muted),
-            ]));
-        }
-    }
-
-    let hint = "  ↑↓ navigate · Enter select · Esc cancel  ";
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled(hint, theme.muted)));
-
-    f.render_widget(Paragraph::new(lines), list_area);
-    draw_picker_border(f, box_area, theme);
-}
-
-fn draw_diff_scrubber(f: &mut ratatui::Frame, area: Rect, app: &AppState) {
-    let theme = &app.theme;
-    let width = area.width.min(96);
-    let total = app.scrubber_hunks.len() as u16;
-    let height = (total * 4 + 8).min(area.height.saturating_sub(2));
-    let x = area.x + (area.width.saturating_sub(width)) / 2;
-    let y = area.y + (area.height.saturating_sub(height)) / 2;
-    let box_area = Rect {
-        x,
-        y,
-        width,
-        height,
-    };
-    if box_area.height < 6 {
-        return;
-    }
-
-    let bar = Rect {
-        x: box_area.x,
-        y: box_area.y,
-        width: box_area.width,
-        height: 1,
-    };
-    let title = format!(
-        "  Edit Scrubber — {} ({} hunks)  ",
-        app.scrubber_path, total
-    );
-    f.render_widget(
-        Paragraph::new(title).style(theme.accent.add_modifier(Modifier::BOLD)),
-        bar,
-    );
-
-    let list_area = Rect {
-        x: box_area.x,
-        y: box_area.y + 1,
-        width: box_area.width,
-        height: box_area.height.saturating_sub(2),
-    };
-
-    let mut lines: Vec<Line<'static>> = Vec::new();
-    let active = app
-        .scrubber_selection
-        .min(app.scrubber_hunks.len().saturating_sub(1));
-    for (i, hunk) in app.scrubber_hunks.iter().enumerate() {
-        let is_active = i == active;
-        let accepted = app.scrubber_accepted.get(i).copied().unwrap_or(true);
-        let marker = if is_active { "▸ " } else { "  " };
-        let state = if accepted { "[✓]" } else { "[ ]" };
-        let header = format!(
-            "{marker}{state} hunk {} of {} · line {}",
-            i + 1,
-            total,
-            hunk.line_no
-        );
-        let header_style = Style::default()
-            .fg(theme.body_fg)
-            .add_modifier(if is_active {
-                Modifier::BOLD
-            } else {
-                Modifier::empty()
-            });
-        lines.push(Line::from(Span::styled(header, header_style)));
-        for before in &hunk.before_lines {
-            lines.push(Line::from(vec![
-                Span::styled(
-                    "    - ",
-                    Style::default().fg(crate::tui::theme::Palette::RED),
-                ),
-                Span::styled(
-                    before.clone(),
-                    Style::default().fg(crate::tui::theme::Palette::RED),
-                ),
-            ]));
-        }
-        for after in &hunk.after_lines {
-            lines.push(Line::from(vec![
-                Span::styled(
-                    "    + ",
-                    Style::default().fg(crate::tui::theme::Palette::GREEN),
-                ),
-                Span::styled(
-                    after.clone(),
-                    Style::default().fg(crate::tui::theme::Palette::GREEN),
-                ),
-            ]));
-        }
-        lines.push(Line::from(""));
-    }
-
-    let hint = "  ↑↓ navigate · y/n toggle · Y all · N none · Enter apply · Esc cancel  ";
-    lines.push(Line::from(Span::styled(hint, theme.muted)));
-
-    f.render_widget(Paragraph::new(lines), list_area);
-    draw_picker_border(f, box_area, theme);
-}
-
-/// Simple border drawn with box-drawing characters.
-fn draw_picker_border(f: &mut ratatui::Frame, r: Rect, theme: &Theme) {
-    let style = theme.border;
-    // Top
-    if r.height > 0 {
-        let top = "─".repeat(r.width as usize);
-        f.render_widget(
-            Paragraph::new(top).style(style),
-            Rect {
-                x: r.x,
-                y: r.y,
-                width: r.width,
-                height: 1,
-            },
-        );
-    }
-    // Bottom
-    if r.height > 1 {
-        let bot = "─".repeat(r.width as usize);
-        f.render_widget(
-            Paragraph::new(bot).style(style),
-            Rect {
-                x: r.x,
-                y: r.y + r.height - 1,
-                width: r.width,
-                height: 1,
-            },
-        );
-    }
-    // Left edge (corners overlap — good enough for a 1px line)
-    if r.height > 2 {
-        let left: Vec<Line<'static>> = (1..r.height - 1)
-            .map(|_| Line::from(Span::styled("│", style)))
-            .collect();
-        f.render_widget(
-            Paragraph::new(left),
-            Rect {
-                x: r.x,
-                y: r.y + 1,
-                width: 1,
-                height: r.height - 2,
-            },
-        );
-        let right: Vec<Line<'static>> = (1..r.height - 1)
-            .map(|_| Line::from(Span::styled("│", style)))
-            .collect();
-        f.render_widget(
-            Paragraph::new(right),
-            Rect {
-                x: r.x + r.width - 1,
-                y: r.y + 1,
-                width: 1,
-                height: r.height - 2,
-            },
-        );
-    }
-}
-
-fn init_terminal() -> Result<Terminal<ratatui::backend::CrosstermBackend<std::io::Stdout>>> {
-    ratatui::crossterm::terminal::enable_raw_mode()?;
-    let mut stdout = std::io::stdout();
-    ratatui::crossterm::execute!(stdout, ratatui::crossterm::terminal::EnterAlternateScreen)?;
-    let backend = ratatui::backend::CrosstermBackend::new(stdout);
-    let terminal = Terminal::new(backend)?;
-    Ok(terminal)
-}
-
-fn restore_terminal() -> Result<()> {
-    let _ = ratatui::crossterm::terminal::disable_raw_mode();
-    ratatui::crossterm::execute!(
-        std::io::stdout(),
-        ratatui::crossterm::terminal::LeaveAlternateScreen,
-    )?;
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn stream_batching_caps_stream_redraws_to_frame_budget() {
-        let start = Instant::now();
-
-        assert_eq!(
-            render_wake_for_stream_batch(start, start + Duration::from_millis(1), false),
-            RenderWake::Wait(Duration::from_millis(15))
-        );
-    }
-
-    #[test]
-    fn input_events_render_immediately() {
-        let start = Instant::now();
-
-        assert_eq!(
-            render_wake_for_stream_batch(start, start + Duration::from_millis(1), true),
-            RenderWake::Now
-        );
-    }
-
-    #[test]
-    fn model_command_is_available_and_deferred_mid_turn() {
-        let command = SLASH_COMMANDS
-            .iter()
-            .find(|cmd| cmd.name == "model")
-            .expect("model slash command");
-
-        assert!(!command.mid_turn_safe);
-        assert_eq!(filter_commands("mod")[0].name, "model");
-    }
-
-    #[test]
-    fn build_provider_picker_entries_lists_default_first_then_named_sorted() {
-        use crate::config::{Config, ProviderConfig};
-        use std::collections::HashMap;
-
-        let mut providers = HashMap::new();
-        let mut local = ProviderConfig::default();
-        local.base_url = "http://localhost:11434/v1".into();
-        local.model = "qwen2.5".into();
-        providers.insert("local".into(), local);
-        let mut alpha = ProviderConfig::default();
-        alpha.base_url = "https://alpha.example".into();
-        alpha.model = "alpha-1".into();
-        providers.insert("alpha".into(), alpha);
-
-        let mut config = Config::default();
-        config.provider.base_url = "https://openrouter.ai/api/v1".into();
-        config.provider.model = "deepseek/v4".into();
-        config.providers = providers;
-
-        let entries = build_provider_picker_entries(&config);
-        assert_eq!(entries.len(), 3);
-        assert!(entries[0].name.is_none());
-        assert_eq!(entries[0].model, "deepseek/v4");
-        assert_eq!(entries[1].name.as_deref(), Some("alpha"));
-        assert_eq!(entries[2].name.as_deref(), Some("local"));
-    }
-
-    #[test]
-    fn provider_command_is_available_and_deferred_mid_turn() {
-        let command = SLASH_COMMANDS
-            .iter()
-            .find(|cmd| cmd.name == "provider")
-            .expect("provider slash command");
-
-        assert!(!command.mid_turn_safe);
-        assert_eq!(filter_commands("prov")[0].name, "provider");
-    }
-
-    #[test]
-    fn format_provider_list_marks_active_profile_and_lists_named() {
-        use crate::config::{Config, ProviderConfig};
-        use std::collections::HashMap;
-
-        let mut providers = HashMap::new();
-        let mut local = ProviderConfig::default();
-        local.base_url = "http://localhost:11434/v1".into();
-        local.model = "qwen2.5".into();
-        providers.insert("local".into(), local);
-
-        let mut config = Config::default();
-        config.provider.base_url = "https://openrouter.ai/api/v1".into();
-        config.provider.model = "deepseek/v4".into();
-        config.providers = providers;
-
-        let active_default = format_provider_list(&config, None);
-        assert!(active_default.contains("* default · https://openrouter.ai/api/v1 · deepseek/v4"));
-        assert!(active_default.contains("  local · http://localhost:11434/v1 · qwen2.5"));
-
-        let active_local = format_provider_list(&config, Some("local"));
-        assert!(active_local.contains("  default · https://openrouter.ai/api/v1"));
-        assert!(active_local.contains("* local · http://localhost:11434/v1"));
-    }
-
-    #[test]
-    fn format_provider_list_handles_no_named_profiles() {
-        use crate::config::Config;
-        let config = Config::default();
-        let report = format_provider_list(&config, None);
-        assert!(report.contains("* default"));
-        assert!(report.contains("(no named [providers.<name>] profiles configured)"));
-    }
-
-    #[test]
-    fn format_model_list_marks_active_model() {
-        let models = vec![
-            crate::provider::catalog::ModelInfo {
-                id: "model-a".into(),
-                display_name: "Model A".into(),
-                context_length: 1_000,
-                pricing: None,
-                features: crate::provider::catalog::ModelFeatures {
-                    tools: true,
-                    vision: false,
-                    json_mode: true,
-                    reasoning: false,
-                },
-                top_provider: None,
-            },
-            crate::provider::catalog::ModelInfo {
-                id: "model-b".into(),
-                display_name: "Model B".into(),
-                context_length: 0,
-                pricing: None,
-                features: crate::provider::catalog::ModelFeatures::default(),
-                top_provider: None,
-            },
-        ];
-
-        let report = format_model_list("model-a", &models);
-
-        assert!(report.contains("* model-a · ctx 1,000 · tools,json"));
-        assert!(report.contains("  model-b · ctx unknown"));
-        assert!(report.contains("Use /model <id> to switch."));
-    }
-}
